refactor(referral): streamline referral feature integration and error handling
- Updated middleware to return a 404 response for referral routes when the feature is disabled, enhancing user feedback.
- Introduced a new utility function, isReferralEnabled, to centralize feature flag checks across components and API routes.
- Removed the deprecated referralFeature module, consolidating referral logic within the utils for better maintainability.
- Adjusted environment configuration to ensure consistent handling of the referral feature flag.
