--- app/lib/referralFeature.ts
@@ -1,6 +1,0 @@
-import config from "./config";
-
-/** Client feature flag: referral program. */
-export function isReferralEnabled(): boolean {
-  return config.referralEnabled;
-}
